# Add benchmark comparing performance of this package with node:zlib.crc32